requirements/dev.txt
-r prod.txt

coverage>=4.4.1
coveralls>=1.2.0
flake8>=3.5.0
pytest>=3.2.3
pytest-django>=3.1.2
pytest-mock>=1.6.3
pytest-pythonpath>=0.7.1
tox>=2.9.1
